backend/imageDetection/requirements.txt
fastapi
uvicorn
opencv-python
pillow
numpy
python-multipart
# PyTorch
torch
torchvision
torchaudio